🧹 [code health improvement] Rename utlis directory to utils
