+++ backend/controllers/userController.js
@@ -1,8 +1,8 @@
-const ErrorHandler = require("../utlis/errorhandler");
+const ErrorHandler = require("../utils/errorhandler");
 const catchAsyncErrors = require("../middleware/catchAsyncErrors");
 const User = require("../models/userModel");
-const sendToken = require("../utlis/jwtToken");
-const sendEmail = require("../utlis/sendEmail")
+const sendToken = require("../utils/jwtToken");
+const sendEmail = require("../utils/sendEmail")
 const crypto = require("crypto")
 const cloudinary = require("cloudinary")
