+++ backend/middleware/error.js
@@ -1,4 +1,4 @@
-const ErrorHandler = require("../utlis/errorhandler")
+const ErrorHandler = require("../utils/errorhandler")
 module.exports = (err, req, res, next) => {
     err.statusCode = err.statusCode || 500;
     err.message = err.message || "internal server error"
