+++ backend/middleware/validate.js
@@ -1,5 +1,5 @@
 const { z } = require("zod");
-const ErrorHandler = require("../utlis/errorhandler");
+const ErrorHandler = require("../utils/errorhandler");
 
 const validate = (schema) => (req, res, next) => {
     const result = schema.safeParse(req.body);
