+++ backend/tests/unit/apifeatures.test.js
@@ -1,4 +1,4 @@
-const Apifeatures = require('../../utlis/apifeatures');
+const Apifeatures = require('../../utils/apifeatures');
 
 describe('Apifeatures Class', () => {
     let mockQuery;
