+++ backend/tests/unit/imageHandler.test.js
@@ -1,5 +1,5 @@
 
-const { processImagesUpdate } = require('../../utlis/imageHandler');
+const { processImagesUpdate } = require('../../utils/imageHandler');
 const cloudinary = require('cloudinary');
 
 jest.mock('cloudinary', () => ({
