+++ backend/utils/imageHandler.js
